Get Ikiwiki stuff working minimally.
* examples/ikiwiki/IkiWiki/Plugin/muse.pm: Move here from
  examples/ikiwiki/muse.  Encode/decode UTF-8 properly.  Take pains to
  remove the temp file.  Die on errors.  Remove test sub.

* examples/ikiwiki/muse-init-project.el: Remove server stuff.

* lisp/muse-ikiwiki.el (muse-ikiwiki-publish-file): Set font-lock-verbose
  to nil to avoid spurious messages when publishing.  Silence the "Wrote
  file" message.

* lisp/muse-publish.el (muse-batch-publish-files): Set font-lock-verbose
  to nil to avoid spurious messages when publishing.

* lisp/muse.el (muse-write-file): Take NOMESSAGE argument.

* texi/muse.texi (Ikiwiki): Add instructions for using
  IkiWiki::Plugin::muse.

+++ examples/ikiwiki/IkiWiki/Plugin/muse.pm
@@ -4,8 +4,9 @@
 # License: GPLv2 or later
 #
 # In your ikiwiki.setup file, set the muse_init option to the location
-# of the init file for Muse.  Some examples provided in this directory
-# are muse-init-simple.el and muse-init-project.el.
+# of the init file for Muse.  Some examples provided in the
+# examples/ikiwiki directory are muse-init-simple.el and
+# muse-init-project.el.
 
 package IkiWiki::Plugin::muse;
 
@@ -21,7 +22,7 @@ sub import {
 
 sub htmlize (@) {
     my %params=@_;
-    my $content = decode_utf8(encode_utf8($params{content}));
+    my $content = encode_utf8($params{content});
     my $qname = $params{page};
     $qname =~ s/"/\\"/g;
 
@@ -32,22 +33,21 @@ sub htmlize (@) {
     $qfile =~ s/"/\\"/g;
     eval {
         system qw( emacs -q --no-site-file -batch -l ),
-               $config{muse_init}, '--eval',
-               qq{(muse-ikiwiki-publish-file "$qfile" "$qname")};
+          $config{muse_init}, '--eval',
+            qq{(muse-ikiwiki-publish-file "$qfile" "$qname")};
         {
             open my $ifh, '<', $filename;
             local $/;
             $content = <$ifh>;
             close $ifh;
         }
+        unlink $filename;
     };
-    unlink $filename;
-    return $content;
-}
-
-sub test {
-    print htmlize(content => "<example>\nHello\n</example>\n\nParagraph.\n",
-                  page => "some_page.muse");
+    if ($@) {
+        unlink $filename;
+        die $@;
+    }
+    return decode_utf8($content);
 }
 
 1
